pyproject.toml
[tool.poetry]
name = "favorite-languages-for-aoc"
version = "0.1.0"
description = ""
authors = ["Ian Wolterstorff <wolterstorff.i@northeastern.edu>"]

[tool.poetry.dependencies]
python = "^3.8"
PyGithub = "^1.55"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
